+++ src/backends/meta-monitor.h
@@ -160,8 +160,8 @@ void meta_monitor_calculate_crtc_pos (MetaMonitor         *monitor,
                                       int                 *out_x,
                                       int                 *out_y);
 
-int meta_monitor_calculate_mode_scale (MetaMonitor     *monitor,
-                                       MetaMonitorMode *monitor_mode);
+float meta_monitor_calculate_mode_scale (MetaMonitor     *monitor,
+                                         MetaMonitorMode *monitor_mode);
 
 MetaMonitorModeSpec * meta_monitor_mode_get_spec (MetaMonitorMode *monitor_mode);
