+++ src/backends/x11/meta-monitor-manager-xrandr.c
@@ -1791,7 +1791,7 @@ meta_monitor_manager_xrandr_is_transform_handled (MetaMonitorManager  *manager,
   return TRUE;
 }
 
-static int
+static float
 meta_monitor_manager_xrandr_calculate_monitor_mode_scale (MetaMonitorManager *manager,
                                                           MetaMonitor        *monitor,
                                                           MetaMonitorMode    *monitor_mode)
